fix forced-checkout
In order for checkout to overwrite dirty files, the repository's working
directory must be set to be the same as the checkout target. This is
because the working directory is used as the working tree which is
compared to the given revision, e.g. to determine if a file is dirty or
not.